RepairA

+++ src/RePair/Adrian/array.h
@@ -34,22 +34,22 @@ Chile. Blanco Encalada 2120, Santiago, Chile. gnavarro@dcc.uchile.cl
 namespace repair_gn{
 
     typedef struct
-    { int *pairs; // identifiers
-        int maxsize;
-        int size;
-        int fst; // first of circular array
+    {   int64_t *pairs; // identifiers
+        int64_t maxsize;
+        int64_t size;
+        int64_t fst; // first of circular array
         float factor;
-        int minsize;
+        int64_t minsize;
         void *Rec; // records
     } Tarray;
 
 // contents can be accessed as A.pairs[0..A.size-1]
 
-    int insertArray (Tarray *A, int pair); // inserts pair in A, returns pos
+    int64_t insertArray (Tarray *A, int64_t pair); // inserts pair in A, returns pos
 
     void deleteArray (Tarray *A); // deletes last cell in A
 
-    Tarray createArray (void *Rec, float factor, int minsize); // creates empty array
+    Tarray createArray (void *Rec, float factor, int64_t minsize); // creates empty array
 
     void destroyArray (Tarray *A); // destroys A
 }
